Refactored the sonar setup
