+++ src/app/chat/component/client-info-panel/mock-translate.mock.ts
